# perf: Optimize NEXT_INST macro by reordering memory loads for better performance