--- a/src/machine/asm/execute_x64.S
+++ b/src/machine/asm/execute_x64.S
@@ -374,11 +374,11 @@
  */
 #define NEXT_INST \
   movq (INST_ARGS), %rcx; \
+  movq (INST_PC), TEMP1; \
   addq $16, INST_ARGS; \
+  addq $16, INST_PC; \
   movzbl %ch, RDd_RS2sd; \
   sar $32, %rcx; \
-  movq (INST_PC), TEMP1; \
-  addq $16, INST_PC; \
   jmp *TEMP1
 
 #define NEXT_INST_V2 \