@mohanson 2025-03-12
Confirmed that the performance has been significantly improved.

```
interpret secp256k1_bench via assembly
                        time:   [4.5208 ms 4.5343 ms 4.5484 ms]
                        change: [-1.9489% -1.4873% -1.0056%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

interpret secp256k1_bench via assembly mop
                        time:   [4.4476 ms 4.4651 ms 4.4878 ms]
                        change: [-2.4842% -1.9465% -1.3296%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

interpret secp256k1_bench via assembly mop (memoized decoder)
                        time:   [3.7501 ms 3.7695 ms 3.7943 ms]
                        change: [-4.0137% -3.3920% -2.7504%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe

interpret secp256k1_bench via assembly mop (memoized dynamic length decoder)
                        time:   [3.1980 ms 3.2061 ms 3.2149 ms]
                        change: [-2.9800% -2.3961% -1.8728%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
```